sparx-times-table-solver-main/requirements.txt
customtkinter>=5.2.0
easyocr>=1.7.0
pyautogui>=0.9.54
Pillow>=10.0.0
numpy>=1.24.0
sympy>=1.12
opencv-python>=4.8.0
